[Nexmark] Extract PersonGenerator, StringsGenerator, LongGenerator from Generator
